feat: updating cart counter and products information into minicart
