+++ src/components/product-card.tsx
@@ -2,6 +2,7 @@ import { ShoppingBag } from "lucide-react";
 import Link from "next/link";
 import { Product } from "../styles/components/product-card";
 import Image from "next/image";
+import { formatPrice } from "../utils/formatPrice";
 
 export interface ProductType {
   id: string;
@@ -11,11 +12,6 @@ export interface ProductType {
 }
 
 export default function ProductCard({ id, name, imageUrl, price }: ProductType) {
-  const formatPrice = new Intl.NumberFormat('pt-BR', {
-    style: 'currency',
-    currency: 'BRL'
-  }).format(price / 100)
-
   return (
     <Link href={`/product/${id}`} prefetch={false}>
       <Product className="keen-slider__slide">
@@ -24,7 +20,7 @@ export default function ProductCard({ id, name, imageUrl, price }: ProductType) 
         <footer>
           <section>
             <strong>{name}</strong>
-            <span>{formatPrice}</span>
+            <span>{formatPrice(price)}</span>
           </section>
 
           <button type="button">
