+++ src/pages/_app.tsx
@@ -1,22 +1,17 @@
 import type { AppProps } from 'next/app'
 import { globalStyles } from '../styles/global'
 
-import logoImg from '../assets/logo.svg'
-import { CartButton, Container, Header } from '../styles/pages/app';
-import Image from 'next/image';
-import Link from 'next/link';
-import { ShoppingBag } from 'lucide-react';
+import { Container } from '../styles/pages/app';
 import Minicart from '../components/minicart';
 import { useState } from 'react';
-import { CartProvider, useShoppingCart } from 'use-shopping-cart';
+import { CartProvider } from 'use-shopping-cart';
+import Header from '../components/header';
 
 globalStyles();
 
 export default function App({ Component, pageProps }: AppProps) {
   const stripeKey = process.env.STRIPE_PUBLIC_KEY || ""
   const [showMinicart, setShowMinicart] = useState(false)
-
-  const cartCounter = 0
 
   function handleOpenMinicart() {
     setShowMinicart(true)
@@ -32,24 +27,16 @@ export default function App({ Component, pageProps }: AppProps) {
         cartMode="checkout-session"
         stripe={stripeKey}
         currency="BRL"
-        loading={<p>Muita calma nessa hora...</p>}
+        loading={<p>Loading...</p>}
         shouldPersist={true}
       >
-        <Header>
-          <Link href="/" prefetch={false}>
-            <Image src={logoImg} alt="" />
-          </Link>
-
-          <CartButton type="button" onClick={handleOpenMinicart}>
-            {cartCounter > 0 && (
-              <span>1</span>
-            )}
-            <ShoppingBag size={24} />
-          </CartButton>
-        </Header>
-
-        <Minicart isActive={showMinicart} onClose={handleCloseMinicart} />
-
+        <Header
+          onOpenMinicart={handleOpenMinicart}
+        />
+        <Minicart
+          isActive={showMinicart}
+          onClose={handleCloseMinicart}
+        />
         <Component {...pageProps} />
       </CartProvider>
     </Container>
